Add files via upload
This program uses JavaFX as the GUI. This had some unique challenges of its own as the environment was largely static and the objects needed constant updating to feel like the everyday apps we use today. The solution I found was to implement the updates into the various events that were coded into the game such as pressing images and buttons as that was the only way to repeat certain lines of code.